refactor: Outsourse Home page codes to a separate file
